Большой пакет ТГ-изменений.

+++ code/modules/power/gravitygenerator.dm
@@ -69,7 +69,6 @@ var/const/GRAV_NEEDS_WRENCH = 3
 /obj/machinery/gravity_generator/part/Destroy()
 	if(main_part)
 		qdel(main_part)
-		return QDEL_HINT_LETMELIVE
 	set_broken()
 	return ..()
 
@@ -98,7 +97,8 @@ var/const/GRAV_NEEDS_WRENCH = 3
 // Generator which spawns with the station.
 //
 
-/obj/machinery/gravity_generator/main/station/initialize()
+/obj/machinery/gravity_generator/main/station/Initialize()
+	..()
 	setup_parts()
 	middle.add_overlay("activated")
 	update_list()
@@ -108,10 +108,6 @@ var/const/GRAV_NEEDS_WRENCH = 3
 //
 /obj/machinery/gravity_generator/main/station/admin
 	use_power = 0
-
-/obj/machinery/gravity_generator/main/station/admin/New()
-	..()
-	initialize()
 
 //
 // Main Generator with the main code
@@ -140,7 +136,8 @@ var/const/GRAV_NEEDS_WRENCH = 3
 	update_list()
 	for(var/obj/machinery/gravity_generator/part/O in parts)
 		O.main_part = null
-		qdel(O)
+		if(!QDESTROYING(O))
+			qdel(O)
 	return ..()
 
 /obj/machinery/gravity_generator/main/proc/setup_parts()
